Allow directly setting an array as the value of a config option
Permissions cannot be validated for this action because if an array is passed, it's assumed that the script that passes the array does the verification steps. This simplifies implementation, because if a situation warrants setting an array as value, there is likely not an entry for the key that it tries to set in config.php, hence a custom way of validating permissions is required.

This is only used for webhooks.