Add initial permission check for apply-config.php
The script will still refuse to apply config updates if a user tries changing a setting in a domain they do not have access to, even if they have access to another admin domain and thus passes the committed check. This is handled in config.php. The committed check is a blanket permissions check to determine whether a user has any admin permissions at all, and should thus be more robust for users who do not have access to any admin domains.