Update POI geofence I18N to reflect named geofences patch
After that patch, the user selects a geofence from a list of geofences (as set up on the geofences page in the administration interface) rather than inputting a list of coordinates directly into this field. As such, explaining how to set up a geofence with coordinate pairs is no longer in scope for helping users configure this setting - explanations should be made on the geofences list page or in the documentation instead.