Field research matching bugfix
Fixed a bug where webhooks would not trigger unless *all* objectives/rewards on the webhook matched. If there were more than one objective or reward filter on the webhook, only one of them could ever match, so the hook trigger would be skipped. The fix ensures that if any match is found, a $match variable is set to true, and cannot be set back to false if another filter is not matching. The filtering mode (whitelist or blacklist) is then evaluated against $match to ensure that any, rather than all, filters must be satisfied in order to trigger the webhook.