Allow research parameter wildcards for webhooks
This allows adminstrators to set up webhooks that match a particular type of objective and/or reward without having to restrict it to also match a given set of parameters. E.g. if you wanted to have a webhook for potion rewards, you'd have to specify the number of potions as well, and reported research that did not exactly match the specified number of potions would not trigger the webhook. Now, it would be possible to disable the "quantity" parameter of the reward, and the webhook would be triggered if potions were reported as a reward, regardless of the quantity.