Fix reporting research not unbinding event handlers
When reporting research, #update-poi-details is displayed, not #poi-details, causing the close event to not be considered valid. Updated to use currentMarker instead, which is initialized as -1 on load, set to an ID when opening a POI, and never reset to -1 anywhere except in the same if block as this patch is applied.