Limit geolocation zoom to reasonable level
Geolocation on maps with Leaflet would zoom in to level 20, the maximum permitted, when using high-accuracy devices, which could cause relevant POIs to be positioned off screen. Limiting the maximum zoom level to 18 for geolocation prevents this issue and lets users see a greater portion of the map when clicking the geolocation button.