Navigation links to map on POI list
Instead of pointing to the default off-site navigation provider, the coordinate pair for each POI in the POI list will now pan the FreeField map page to the location of the POI on the map. Furthermore, the "current research" icons in the same table can now be clicked to open POI details for the specified POI so that research can be submitted for it.