Support custom icon sets
Good for developing new icon sets and for testing it internally before hopefully submitting them to the repository for others to enjoy. This could be used for e.g. species sets in a different language, custom art drawn for a community, etc. Admins are of course responsible for whatever they put here.