Fix old research resurfacing when doing actions on POI
Old research was stored in the database, but flagged as out-of-date using the last_updated column in the database. When performing non-research updates, such as evil reports and renames of POIs, this field would be updated, falsely setting the research to an incorrect state. Fixed by manually clearing the research state if it contains an outdated non-unknown task.